Add-Articles-ErrorHandling-And-Conditions

+++ Services/Articles/ArticleServices.cs
@@ -1,14 +1,67 @@
 ﻿using GData.DTOs.ArticlesDTO;
 using GData.Entity;
+using GData.Exceptions;
 using GData.Repositories.Articles;
 using GData.Repositories.Users;
+using GData.Services.Users;
 
 namespace GData.Services.Articles
 {
-    public class ArticleServices(IArticleRepository articleRepository) : IArticleServices
+    public class ArticleServices(IArticleRepository articleRepository, IAuthServices authServices, ArticlesExceptionList articlesExceptionList) : IArticleServices
     {
         public async Task<Article> CreateArticleService(Guid creatorId, ArticleDTO request)
         {
+
+            var creator = await authServices.GetUserByIdService(creatorId);
+
+            if (creator is null)
+            {
+
+                return await articlesExceptionList.ArticleCreatorDoesNotExist();
+
+            }
+
+            if (creator.UserRole != 0)
+            {
+
+                return await articlesExceptionList.InvalidUser();
+
+            }
+
+            if (creator.IsEmailConfirmed == false)
+            {
+
+                return await articlesExceptionList.UnverifiedUserEmail();
+
+            }
+
+            if (string.IsNullOrWhiteSpace(request.Content) || string.IsNullOrWhiteSpace(request.Title) || string.IsNullOrWhiteSpace(request.Author))
+            {
+
+                return await articlesExceptionList.NoDataProvidedForArticle();
+
+            }
+
+            if (string.IsNullOrWhiteSpace(request.Content) && string.IsNullOrWhiteSpace(request.Title) && string.IsNullOrWhiteSpace(request.Author))
+            {
+
+                return await articlesExceptionList.NoDataProvidedForArticle();
+
+            }
+
+            if (request.Content.Length < 4 || request.Title.Length < 4 || request.Author.Length < 4)
+            {
+
+                return await articlesExceptionList.BadDataFormat();
+
+            }
+
+            if (request.Content.Length < 4 && request.Title.Length < 4 && request.Author.Length < 4)
+            {
+
+                return await articlesExceptionList.BadDataFormat();
+
+            }
 
             var article = new Article()
             {
@@ -26,21 +79,70 @@ namespace GData.Services.Articles
 
         }
 
-        public async Task<Article> DeleteArticleService(Guid creatorId, Guid Id)
+        public async Task<Article> DeleteArticleService(Guid Id)
         {
 
             var article = await GetArticleByIdService(Id);
 
-            var result=await articleRepository.DeleteArticle(article);
+            if (article is null)
+            {
+
+                return await articlesExceptionList.ArticleNotFound();
+
+            }
+
+            var result = await articleRepository.DeleteArticle(article);
 
             return result;
 
         }
 
-        public async Task<Article> EditArticleService(Guid creatorId, Guid Id, ArticleDTO request)
+        public async Task<Article> EditArticleService(Guid Id, ArticleDTO request)
         {
 
             var article = await GetArticleByIdService(Id);
+
+            if (article is null)
+            {
+
+                return await articlesExceptionList.ArticleNotFound();
+
+            }
+
+            if (article.ArticleCreator is null)
+            {
+
+                return await articlesExceptionList.EditArticleCreatorDoesNotExist();
+
+            }
+
+            if (request.Content.Length < 4 && request.Title.Length < 4 && request.Author.Length < 4)
+            {
+
+                return await articlesExceptionList.BadDataFormat();
+
+            }
+
+            if (request.Content.Length < 4 || request.Title.Length < 4 || request.Author.Length < 4)
+            {
+
+                return await articlesExceptionList.BadDataFormat();
+
+            }
+
+            if (string.IsNullOrWhiteSpace(request.Author)||string.IsNullOrWhiteSpace(request.Content)||string.IsNullOrWhiteSpace(request.Title))
+            {
+
+                return await articlesExceptionList.NoDataProvidedForArticle();
+
+            }
+
+            if (string.IsNullOrWhiteSpace(request.Author) && string.IsNullOrWhiteSpace(request.Content) && string.IsNullOrWhiteSpace(request.Title))
+            {
+
+                return await articlesExceptionList.NoDataProvidedForArticle();
+
+            }
 
             var result=await articleRepository.EditArticle(article, request);
 
@@ -59,6 +161,14 @@ namespace GData.Services.Articles
         {
 
             var article = await articleRepository.GetArticleById(Id);
+           
+            if(article is null)
+            {
+
+                return await articlesExceptionList.ArticleNotFound();
+
+            }
+
             return article;
 
         }
@@ -67,6 +177,14 @@ namespace GData.Services.Articles
         {
             
             var article= await articleRepository.GetArticleByTitle(title);
+
+            if (article is null)
+            {
+
+                return await articlesExceptionList.ArticleNotFound();
+
+            }
+
             return article;
 
         }
